♻️ with-navbarのリファクタ
